pull profile component into its own file and styling sheet for better maintainability and cleaner component structure
