+++ client/src/components/MenuBar.jsx
@@ -1,16 +1,9 @@
-import { useEffect, useState } from "react";
+import { useState } from "react";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import {
-  faUserPlus,
-  faUserGear,
-  faCircleXmark,
-  faCircle,
-  faMoon,
-  faXmark,
-  faHeartCrack,
-} from "@fortawesome/free-solid-svg-icons";
+import { faUserPlus, faUserGear } from "@fortawesome/free-solid-svg-icons";
 import { v4 } from "uuid";
 import ChatRow from "./subComponents/ChatRow";
+import Profile from "./subComponents/Profile";
 
 function MenuBar(props) {
   const [showProfile, setShowProfile] = useState(false);
@@ -52,45 +45,6 @@ function MenuBar(props) {
         <Profile setShowProfile={setShowProfile} user={props.user} />
       ) : null}
     </aside>
-  );
-}
-
-function Profile(props) {
-  return (
-    <div className="ProfileBackdrop">
-      <div className="Profile">
-        <ul className="topBar">
-          <li onClick={() => props.setShowProfile(false)}>
-            <FontAwesomeIcon icon={faCircleXmark} className="iconClose" />
-          </li>
-        </ul>
-
-        <div className="profileInformation">
-          <p className="profileImage">
-            <FontAwesomeIcon icon={faCircle} style={{ fontSize: "150px" }} />
-          </p>
-          <p>{props.user.fullName}</p>
-          <p>({props.user.username})</p>
-        </div>
-
-        <nav className="profileButtons">
-          <ul className="buttonList">
-            <li>
-              <FontAwesomeIcon className="buttonIcon" icon={faMoon} />
-              <button>Toggle Dark Mode</button>
-            </li>
-            <li>
-              <FontAwesomeIcon className="buttonIcon" icon={faXmark} />
-              <button>Log Out</button>
-            </li>
-            <li>
-              <FontAwesomeIcon className="buttonIcon" icon={faHeartCrack} />
-              <button>Report A Problem</button>
-            </li>
-          </ul>
-        </nav>
-      </div>
-    </div>
   );
 }
