+++ client/src/scss/_menuBar.scss
@@ -34,91 +34,3 @@
     height: calc(100vh - 50px);
   }
 }
-
-.ProfileBackdrop {
-  height: 100vh;
-  width: 100vw;
-  background-color: rgb(0, 0, 0, 0.4);
-
-  position: absolute;
-  top: 0;
-  bottom: left;
-
-  .Profile {
-    // Matches menuBar as to not let it spill underneath
-    width: clamp(320px, 100vw, 500px);
-    height: 100vh;
-
-    position: absolute;
-    top: 0;
-    bottom: 0;
-
-    padding: 0 10px 0 10px;
-    background-color: v.$theme-main-bk-secondary;
-
-    display: flex;
-    flex-direction: column;
-    gap: 20px;
-    .topBar {
-      height: 50px;
-      padding-right: 10px;
-
-      display: flex;
-      justify-content: flex-end;
-      align-items: center;
-      li {
-        list-style: none;
-        .iconClose {
-          color: white;
-          background-color: black;
-          border-radius: 50%;
-          font-size: 25px;
-          cursor: pointer;
-        }
-      }
-    }
-
-    .profileInformation {
-      display: flex;
-      flex-direction: column;
-      align-items: center;
-    }
-
-    .profileButtons {
-      background-color: rgb(214, 216, 218);
-      border-radius: 10px;
-      padding: 10px 10px 10px 10px;
-      .buttonList {
-        display: flex;
-        flex-direction: column;
-        li {
-          padding: 10px;
-          list-style: none;
-          border-radius: 20px;
-
-          display: flex;
-          justify-content: center;
-          align-items: center;
-          .buttonIcon {
-            cursor: pointer;
-            flex: 1;
-            font-size: 22px;
-          }
-          button {
-            cursor: pointer;
-            text-align: left;
-            flex: 5;
-            border: none;
-            background-color: transparent;
-            padding-bottom: 6px;
-            border-bottom: rgba(30, 30, 30, 0.4) solid 1px;
-          }
-        }
-        li:hover {
-          background-color: white;
-          cursor: pointer;
-        }
-      }
-    }
-  }
-}
