+++ client/src/scss/index.scss
@@ -8,6 +8,7 @@
 // SubComponents Styling
 @import "./subComponents/message";
 @import "./subComponents/chatRow";
+@import "./subComponents/profile";
 // Forms styling
 @import "./forms/createAccountForm";
 @import "./forms/newMessageForm";
